requirements.txt
streamlit==1.44.1
pandas==2.2.3
numpy==2.1.3
plotly==6.0.1
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-dotenv==1.0.1
openai==1.75.0
requests==2.32.3